requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.30.0
marshmallow>=3.0,<4.0
Flask-JWT-Extended==4.6.0
Werkzeug>=2.3,<3.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
setuptools
